+++ sansmodels/src/sans/models/c_models/ellipsoid.cpp
@@ -77,11 +77,11 @@ double EllipsoidModel :: operator()(double q) {
 	double vol = 0.0;
 
 	// Loop over radius_a weight points
-	for(int i=0; i<weights_rad_a.size(); i++) {
+	for(size_t i=0; i<weights_rad_a.size(); i++) {
 		dp[1] = weights_rad_a[i].value;
 
 		// Loop over radius_b weight points
-		for(int j=0; j<weights_rad_b.size(); j++) {
+		for(size_t j=0; j<weights_rad_b.size(); j++) {
 			dp[2] = weights_rad_b[j].value;
 			//Un-normalize  by volume
 			sum += weights_rad_a[i].weight
@@ -146,20 +146,20 @@ double EllipsoidModel :: operator()(double qx, double qy) {
 	double vol = 0.0;
 	double pi = 4.0*atan(1.0);
 	// Loop over radius weight points
-	for(int i=0; i<weights_rad_a.size(); i++) {
+	for(size_t i=0; i<weights_rad_a.size(); i++) {
 		dp.radius_a = weights_rad_a[i].value;
 
 
 		// Loop over length weight points
-		for(int j=0; j<weights_rad_b.size(); j++) {
+		for(size_t j=0; j<weights_rad_b.size(); j++) {
 			dp.radius_b = weights_rad_b[j].value;
 
 			// Average over theta distribution
-			for(int k=0; k<weights_theta.size(); k++) {
+			for(size_t k=0; k<weights_theta.size(); k++) {
 				dp.axis_theta = weights_theta[k].value;
 
 				// Average over phi distribution
-				for(int l=0; l<weights_phi.size(); l++) {
+				for(size_t l=0; l<weights_phi.size(); l++) {
 					dp.axis_phi = weights_phi[l].value;
 					//Un-normalize by volume
 					double _ptvalue = weights_rad_a[i].weight
